requirements.txt
# Python dependencies for Blackjack AI Vision
# Training, dataset management, and utilities

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.1.0

# Dataset management
roboflow>=1.1.0

# Computer vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# TensorRT Python API (optional, for Python-based conversion)
# nvidia-tensorrt>=8.6.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Web dashboard
# flask>=2.3.0
# flask-cors>=4.0.0
